+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Support for predict Algos (#1071)
+
 ### Changed
 
 - Use assets embedded in events during sync (#1062)